requirements.txt
matplotlib==3.7.1
numpy==1.26.4
optuna==3.5.0
pandas==2.2.1
PyYAML==6.0.1
rich==13.7.1
scikit_learn==1.3.1
setuptools==69.0.2
torch==2.1.2
transformers==4.25.1
tensorboard